turn and messages have been moved into the avatar from the world and all seems to be working correctly

+++ Splorr.Seafarers.Models/Island.fs
@@ -4,7 +4,7 @@ type Island =
     {
         Name       : string
         VisitCount : uint32 option
-        LastVisit  : uint32 option
+        LastVisit  : float option
         Jobs       : Job list
         Markets    : Map<Commodity, Market>
         Items      : Set<Item>
